requirements.txt
Flask
Flask.CORS
requests
openai
python-dotenv